Andreas8

+++ src/components/CardsList/CardsList.jsx
@@ -6,7 +6,7 @@ import DeleteForeverOutlinedIcon from '@material-ui/icons/DeleteForeverOutlined'
 import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import './css/CardsList.css';
 
-export default function CardsList() {
+export default function CardsList({ createdBy }) {
     const dispatch = useDispatch()
 
     const [cardsList, setCardsList] = useState([]);
@@ -25,7 +25,7 @@ export default function CardsList() {
         fetch(url, options).then((data) =>
             data.json().then((output) => {
                 if (output.status === "success") {
-                    console.log('CARDS', output);
+                    // console.log('CARDS', output);
                     const cards = output.data
                     setCardsList(cards)
                 }
@@ -37,41 +37,45 @@ export default function CardsList() {
         getAllCards()
     }, []);
 
-    const Cards = ({ cards }) => {
+    const Cards = ({ cards, createdBy }) => {
         const cardSet = cards.map((card, i) => {
-            const pass = `http://localhost:8080/greeting-pictures/${card.picture}`
-            const cardID = card._id
 
-            return (
-                <div className="cards-list-wrapper2" key={i}>
-                    <Link key={cardID} to={`/cards/${cardID}`} >
-                        <img
-                            alt='card catalog pic'
-                            className="picture"
-                            src={pass}
-                            // onClick={() => {dispatch(sendPict(picture)); closeModal()}}
-                            style={{ cursor: 'pointer' }}
-                        />
-                    </Link>
-                    <div className="cards-list-btn-container">
-                        <Link
-                            to="#"
-                            onClick={async () => {
-                                const result = await dispatch(deleteCard(card))
-                                if (result.status == 'success') {
-                                    console.log(result);
-                                    console.log('delte operation. Implement useState to up-date a page ');
-                                    getAllCards()
-                                }
-                            }}
-                        >
-                            <p><DeleteForeverOutlinedIcon /></p>
+            if (card.createdBy == createdBy) {
+                const pass = `http://localhost:8080/greeting-pictures/${card.picture}`
+                const cardID = card._id
+
+                return (
+                    <div className="cards-list-wrapper2" key={i}>
+                        <Link key={cardID} to={`/cards/${cardID}`} >
+                            <img
+                                alt='card catalog pic'
+                                className="picture"
+                                src={pass}
+                                // onClick={() => {dispatch(sendPict(picture)); closeModal()}}
+                                style={{ cursor: 'pointer' }}
+                            />
                         </Link>
-                        {/* Has to redirect to card editor with a card */}
-                        <Link to="#" ><p><EditOutlinedIcon /></p></Link>
+                        <div className="cards-list-btn-container">
+                            <Link
+                                to="#"
+                                onClick={async () => {
+
+                                    const result = await dispatch(deleteCard(card))
+                                    if (result.status == 'success') {
+                                        // console.log(result);
+                                        console.log('delte operation. Implement useState to up-date a page ');
+                                        getAllCards()
+                                    }
+                                }}
+                            >
+                                <p><DeleteForeverOutlinedIcon /></p>
+                            </Link>
+                            {/* Has to redirect to card editor with a card */}
+                            <Link to="#" ><p><EditOutlinedIcon /></p></Link>
+                        </div>
                     </div>
-                </div>
-            )
+                )
+            }
         });
 
         return (
@@ -81,12 +85,13 @@ export default function CardsList() {
         );
     };
 
-    console.log(1, cardsList.length);
-
     return (
         <div className="component-cards-list">
             {cardsList.length > 0 ?
-                <Cards cards={cardsList} />
+                <Cards
+                    cards={cardsList}
+                    createdBy={createdBy}
+                />
                 :
                 <h1>There aren't any cars</h1>
             }
